Adds patwahy diversity method to compute similarity between locations
To implement pathway diversity we need to compute the transition
probability between two options. As a proxy of that probability we have
decided to use the cost. Here we use the average pairwise distance
between locations as a proxy for the similarity and the cost of
switching between two options.