Temporarily make cluster_diversity analysis using SPATIAL_GROUPING
When looking to the RME or ReefMod datasets cluster_id has an unique
identifier that does not allow the cluster_diversity analysis (the
result is a vector of zeros). To run the pathway diversity analysis it
temporarily changes the code to use SPATIAL_GROUPING. Later the code
should be able to use it SPATIAL_GROUPING or cluster_id.