Fix new group name and mcda criteria direction
This commit change the name of the group to be used in the diversity
metric from SPATIAL_GROUPING to CB_CALIB_GROUPS, according to the new
name used in the most recent datapackage.
It also change the direction of the seed_geographic_separation from
minimum to maximum, as in this simulation we always want to maximise the
geographic separation.